requirements.txt
# Function dependencies, for example:
# package>=version
python-dateutil
bs4
pandas
sqlalchemy
pymysql